Moved the simd_exec Executors to the policy/simd/kernel path
