+++ include/RAJA/policy/simd.hpp
@@ -30,5 +30,6 @@
 
 #include "RAJA/policy/simd/forall.hpp"
 #include "RAJA/policy/simd/policy.hpp"
+#include "RAJA/policy/simd/kernel/For.hpp"
 
 #endif  // closing endif for header file include guard
